+++ packages/docs/components/TableOfContents/zod-types.tsx
@@ -6,7 +6,7 @@ export const TableOfContents: React.FC = () => {
   return (
     <div>
       <Grid>
-        <TOCItem link="/docs/z-color/z-color">
+        <TOCItem link="/docs/zod-types/z-color">
           <strong>{"zColor()"}</strong>
           <div>A Zod Type for colors</div>
         </TOCItem>
