+++ packages/docs/docs/zod-types/z-color.md
@@ -1,7 +1,7 @@
 ---
 image: /generated/articles-docs-z-color.png
 title: zColor()
-slug: z-color/z-color
+slug: zod-types/z-color
 ---
 
 Using `zColor()`, you can specify in your [schema](/docs/parametrized-rendering) that a string has to be a color.
