+++ packages/example/package.json
@@ -44,7 +44,7 @@
 		"@remotion/shapes": "workspace:*",
 		"@remotion/skia": "workspace:*",
 		"@remotion/three": "workspace:*",
-		"@remotion/z-color": "workspace:*",
+		"@remotion/zod-types": "workspace:*",
 		"@types/express": "^4.17.11",
 		"@types/node": "^16.7.5",
 		"@types/opentype.js": "^1.3.4",
